requirements.txt
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
streamlit==1.51.0
torch==2.9.1
torchvision==0.24.1